Cleanup go contracts source code
