+++ mvm/eos/contracts/dappdemo/dappdemo.abi
@@ -1,0 +1,148 @@
+{
+    "version": "eosio::abi/1.1",
+    "structs": [
+        {
+            "name": "Counter",
+            "base": "",
+            "fields": [
+                {
+                    "name": "id",
+                    "type": "uint64"
+                },
+                {
+                    "name": "count",
+                    "type": "uint64"
+                }
+            ]
+        },
+        {
+            "name": "TxEvent",
+            "base": "",
+            "fields": [
+                {
+                    "name": "nonce",
+                    "type": "uint64"
+                },
+                {
+                    "name": "process",
+                    "type": "uint128"
+                },
+                {
+                    "name": "asset",
+                    "type": "uint128"
+                },
+                {
+                    "name": "members",
+                    "type": "uint128[]"
+                },
+                {
+                    "name": "threshold",
+                    "type": "int32"
+                },
+                {
+                    "name": "amount",
+                    "type": "uint128"
+                },
+                {
+                    "name": "extra",
+                    "type": "bytes"
+                },
+                {
+                    "name": "timestamp",
+                    "type": "uint64"
+                },
+                {
+                    "name": "signature",
+                    "type": "bytes"
+                }
+            ]
+        },
+        {
+            "name": "TxRequest",
+            "base": "",
+            "fields": [
+                {
+                    "name": "nonce",
+                    "type": "uint64"
+                },
+                {
+                    "name": "contract",
+                    "type": "name"
+                },
+                {
+                    "name": "process",
+                    "type": "uint128"
+                },
+                {
+                    "name": "asset",
+                    "type": "uint128"
+                },
+                {
+                    "name": "members",
+                    "type": "uint128[]"
+                },
+                {
+                    "name": "threshold",
+                    "type": "int32"
+                },
+                {
+                    "name": "amount",
+                    "type": "uint128"
+                },
+                {
+                    "name": "extra",
+                    "type": "bytes"
+                },
+                {
+                    "name": "timestamp",
+                    "type": "uint64"
+                }
+            ]
+        },
+        {
+            "name": "onevent",
+            "base": "",
+            "fields": [
+                {
+                    "name": "event",
+                    "type": "TxEvent"
+                }
+            ]
+        }
+    ],
+    "types": [],
+    "actions": [
+        {
+            "name": "onevent",
+            "type": "onevent",
+            "ricardian_contract": ""
+        }
+    ],
+    "tables": [
+        {
+            "name": "counters",
+            "type": "Counter",
+            "index_type": "i64",
+            "key_names": [],
+            "key_types": []
+        },
+        {
+            "name": "txevents",
+            "type": "TxEvent",
+            "index_type": "i64",
+            "key_names": [],
+            "key_types": []
+        },
+        {
+            "name": "txrequests",
+            "type": "TxRequest",
+            "index_type": "i64",
+            "key_names": [],
+            "key_types": []
+        }
+    ],
+    "ricardian_clauses": [],
+    "variants": [],
+    "abi_extensions": [],
+    "error_messages": []
+}
