+++ mvm/eos/contracts/dappdemo/go.mod
@@ -4,4 +4,4 @@ go 1.17
 
 replace github.com/uuosio/chain => /Users/newworld/dev/github/go-chain
 
-require github.com/uuosio/chain v0.1.2
+require github.com/uuosio/chain v0.1.3-0.20211119050052-3f638f6b1849
