--- mvm/eos/contracts/dappdemo/scripts/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-log_cli = 1
-log_cli_level = INFO
