--- mvm/eos/contracts/dappdemo/scripts/test.sh
@@ -1,1 +1,0 @@
-run-uuos -m pytest -s -x test2.py  -k test_mtg_tx
