--- mvm/eos/contracts/dappdemo/utils.go
@@ -1,10 +1,0 @@
-
-package main
-
-import (
-	"github.com/uuosio/chain"
-)
-
-func check(b bool, msg string) {
-	chain.Check(b, msg)
-}
