--- mvm/eos/contracts/dappdemo2/README.md
@@ -1,10 +1,0 @@
-# Building
-
-```bash
-eosio-go build -o dappdemo.wasm .
-```
-
-# Testing
-```
-python3 test.py
-```
