--- mvm/eos/contracts/dappdemo2/actions.go
@@ -1,1 +1,0 @@
-package main
