--- mvm/eos/contracts/dappdemo2/build.sh
@@ -1,2 +1,0 @@
-
-eosio-go build -o dappdemo.wasm .
