--- mvm/eos/contracts/dappdemo2/go.mod
@@ -1,7 +1,0 @@
-module dappdemo
-
-go 1.17
-
-replace github.com/uuosio/chain => /Users/newworld/dev/github/go-chain
-
-require github.com/uuosio/chain v0.1.2
