--- mvm/eos/contracts/dappdemo2/structs.go
@@ -1,7 +1,0 @@
-
-package main
-
-type MyStruct struct {
-	a uint64
-	b uint64
-}
