--- mvm/eos/contracts/dappdemo2/tests/go.mod
@@ -1,5 +1,0 @@
-module test
-
-go 1.17
-
-require github.com/uuosio/chain v0.1.2
