--- mvm/eos/contracts/mtg.xin/structs.go
@@ -1,7 +1,0 @@
-
-package main
-
-type MyStruct struct {
-	a uint64
-	b uint64
-}
